Spring Batch 5 Java 17 remote partitioning with RabbitMQ Configuration Manager Worker Configuration
