--- src/main/java/com/batch/configuration/MasterConfiguration.java
@@ -1,96 +1,0 @@
-package com.batch.configuration;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.amqp.core.AmqpTemplate;
-import org.springframework.amqp.rabbit.connection.ConnectionFactory;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.core.job.builder.JobBuilder;
-import org.springframework.batch.core.launch.support.RunIdIncrementer;
-import org.springframework.batch.core.partition.support.MultiResourcePartitioner;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.integration.partition.RemotePartitioningManagerStepBuilderFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.core.io.Resource;
-import org.springframework.integration.amqp.dsl.Amqp;
-import org.springframework.integration.channel.DirectChannel;
-import org.springframework.integration.channel.QueueChannel;
-import org.springframework.integration.dsl.IntegrationFlow;
-import org.springframework.integration.dsl.MessageChannels;
-
-@RequiredArgsConstructor
-@Profile("master")
-@Configuration
-public class MasterConfiguration {
-
-    private static final int GRID_SIZE = 3;
-
-    private final RemotePartitioningManagerStepBuilderFactory managerStepBuilderFactory;
-
-    /*
-     * Configure outbound flow (requests going to workers)
-     */
-    @Bean
-    public DirectChannel requests() {
-        return MessageChannels.direct().getObject();
-    }
-
-    @Bean
-    public IntegrationFlow outboundFlow(AmqpTemplate amqpTemplate) {
-        return IntegrationFlow.from(requests())
-                .handle(Amqp.outboundAdapter(amqpTemplate).routingKey("requests"))
-                .get();
-    }
-
-    /*
-     * Configure inbound flow (replies coming from workers)
-     */
-    @Bean
-    public QueueChannel replies() {
-        return MessageChannels.queue().getObject();
-    }
-
-    @Bean
-    public IntegrationFlow inboundFlow(ConnectionFactory connectionFactory) {
-        return IntegrationFlow.from(Amqp.inboundAdapter(connectionFactory, "replies"))
-                .channel(replies())
-                .get();
-    }
-
-    @Bean
-    @StepScope
-    public MultiResourcePartitioner partitioner(@Value("#{jobParameters['inputFiles']}") Resource[] resources) {
-        MultiResourcePartitioner partitioner = new MultiResourcePartitioner();
-
-        partitioner.setKeyName("file");
-        partitioner.setResources(resources);
-
-        return partitioner;
-    }
-
-    /*
-     * Configure the manager step
-     */
-    @Bean
-    public Step managerStep(MultiResourcePartitioner partitioner) {
-        return this.managerStepBuilderFactory.get("managerStep")
-                .partitioner("workerStep", partitioner)
-                .gridSize(GRID_SIZE)
-                .outputChannel(requests())
-                .inputChannel(replies())
-                .build();
-    }
-
-    @Bean
-    public Job remotePartitioningJob(JobRepository jobRepository, Step managerStep) {
-        return new JobBuilder("remotePartitioningJob", jobRepository)
-                .start(managerStep)
-                .incrementer(new RunIdIncrementer())
-                .build();
-    }
-
-}
