+++ src/main/resources/application.properties
@@ -1,7 +1,15 @@
+#== RABBIT MQ CONFIGURATION ==#
+spring.rabbitmq.host=localhost
 spring.rabbitmq.username=user
 spring.rabbitmq.password=password
-spring.rabbitmq.host=localhost
+#
+#== POSTGRES CONFIGURATION ==#
 spring.datasource.url=jdbc:postgresql://localhost/postgres
 spring.datasource.username=postgres
 spring.datasource.password=postgres
+#
+#== SPRING BATCH CONFIGURATION ==#
 spring.batch.jdbc.initialize-schema=always
+#
+#== JOB CONFIGURATION ==#
+worker.node.count=3
